+++ src/common/js/index.js
@@ -7,4 +7,5 @@ import './constants.js';
 import './trace.js';
 //
 import './configure.js';
+import './functions.js';
 import './timestampable.js';